organize files by pages
