+++ app/src/pages/not-found/index.js
